Allow latest json-schema version
Besides several additions/fixes, the breaking changes are:
v3.0 dropped support for older Rubies and added tests for Ruby 3+.
v4.0 removed data ivar from JSON::Validator so that multiple validate
call become faster

These don't seem to affect anything on apivore, all tests are green and
apivore doesn't seem to call any of the changed code in v4.0 directly.

I'm also opening up the dependency to `>=` in the gemspec so we don't
need to be updating it for newer versions of json-schema for now.